pyproject.toml
[tool.poetry]
name = "substr"
version = "0.1.0"
description = ""
authors = ["PVPotapov <PVPotapov@greenatom.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.110.2"
uvicorn = "^0.29.0"
python-dotenv = "^1.0.1"
redis = "^5.0.4"
pydantic = "^2.7.1"
aiohttp = "^3.9.5"
logging = "^0.4.9.6"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
